refactor: use internal TCP ping
