+++ libcore/box.go
@@ -13,15 +13,12 @@ import (
 	"libcore/protectserver"
 
 	box "github.com/sagernet/sing-box"
-	"github.com/sagernet/sing-box/adapter"
 	"github.com/sagernet/sing-box/common/conntrack"
-	"github.com/sagernet/sing-box/common/urltest"
 	_ "github.com/sagernet/sing-box/include"
 	boxlog "github.com/sagernet/sing-box/log"
 	"github.com/sagernet/sing-box/option"
 	"github.com/sagernet/sing-box/outbound"
 	E "github.com/sagernet/sing/common/exceptions"
-	N "github.com/sagernet/sing/common/network"
 	"github.com/sagernet/sing/service"
 	"github.com/sagernet/sing/service/pause"
 )
@@ -183,52 +180,16 @@ func (b *BoxInstance) SelectOutbound(tag string) (ok bool) {
 	return false
 }
 
-func UrlTest(i *BoxInstance, link string, timeout int32) (latency int32, err error) {
-	defer device.DeferPanicToError("box.UrlTest", func(err_ error) { err = err_ })
-
-	var router adapter.Router
-	if i == nil {
-		// test current
-		router = mainInstance.Router()
-	} else {
-		router = i.Router()
-	}
-
-	defOutbound, err := router.DefaultOutbound(N.NetworkTCP)
-	if err != nil {
-		return 0, E.Cause(err, "find default outbound")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(timeout)*time.Millisecond)
-	defer cancel()
-
-	chLatency := make(chan uint16, 1)
-	go func() {
-		var t uint16
-		t, err = urltest.URLTest(ctx, link, defOutbound)
-		chLatency <- t
-	}()
-	select {
-	case <-ctx.Done():
-		return 0, E.New("Url test timeout")
-	case t := <-chLatency:
-		if err != nil {
-			return 0, err
-		}
-		return int32(t), nil
-	}
-}
-
 var protectCloser io.Closer
 
 func goServeProtect(start bool) {
 	if protectCloser != nil {
-		protectCloser.Close()
+		_ = protectCloser.Close()
 		protectCloser = nil
 	}
 	if start {
 		protectCloser = protectserver.ServeProtect("protect_path", 0, func(fd int) {
-			intfBox.AutoDetectInterfaceControl(int32(fd))
+			_ = intfBox.AutoDetectInterfaceControl(int32(fd))
 		})
 	}
 }
